[dev.typeparams] Conversion of types2 types to types, compare with existing type
If -G=3 option is provided, then the noder adds a type to most nodes it
generates. For each declaration or expression, we get the types2 type
from the noder.typeInfo type maps. We then transform that types2 type
to a 'types' type using new function typeExpr2() conversion function. (A
bunch of the code in typeExpr2() came from an experimental CL from
Matthew.) Then we set the type on the Node to this converted type and
mark the type as being derived from types2.

To verify correctness, we compare types in miniExpr.SetType() and
Func.SetType(). That is, if we set a type from typechecker.go and there is
already a type on a node derived from types2, we compare the types using
(*Type).Compare. If the types compare identically, then we just keep the
types2-derived type. Otherwise, we print a message and set the type
computed by typechecker.go. So, when we are compiling programs with
-G=3, we are using the types2-derived type for most of the nodes in the
Noder graph for the remainder of the compilation.

I added a new test g3flag.go in the go/test directory. It compiles a list of files
from the test directory with the -G=3 directory and fails if there is any error.
For about five files, it actually does a full "go run -G=3" to test that the
compiled file actually runs without error. This test will be run automatically by
all.bash, thus ensuring that -G=3 won't be broken.

The big areas that I don't deal with yet are forward or circular references among
types, and some cases involving untyped constants.

Change-Id: I1f7d0b209623f8f5e04a124849e3dd283b1aae83