go/ast: improve comment associations in empty function/loop/if bodies
Currently, comments that are placed inside empty function/loop/if bodies are associated with statements after the body, hence statements outside of the scope that the comments are in. With this fix, comments inside empty function/loop/if bodies are associated to the *ast.BlockStmt nodes that represent the empty bodies.

Fixes #39753

Change-Id: I5ff05a3020c8a9cc2ec796667e285b7fbfde0291