cmd/go: add -modversion flag setting main module version on debug.ReadBuildInfo
As discussed in #29228 currently it is required very hacky workarounds to get the version of the built binary.
From my understanding the most used are variations of 'go build -ldflags="-X main.version=${version}"' and reading from the injected variable.

Currently building a binary using "go build" has a hardcoded "(devel)" version for the main module on the data for ReadBuildInfo. The program has to be installed from "go get" or "go install" in order for the version to be added.

This PR adds a "-modversion" to set the value for the main module version.

Change-Id: I8683230902d504d8c386380457a87ec85e5dbcf0
GitHub-Last-Rev: 160c3da46aca8f751c2f494af8c96fc8b07d3801
GitHub-Pull-Request: golang/go#43535