cmd/{asm,compile}: add -minimaldwarf flag
DO NOT SUBMIT
DO NOT REVIEW

Add support for a new compiler command line option -minimaldwarf,
intended to provide just enough DWARF to provide accurate profile
data, but without the bloat/space needed for debugging.

The general strategy is to emit a subprogram DIE for a function only
if we need it as a parent for inlined subroutine DIEs (if no inlining
in a given function, then don't emit a subprogram DIE for it). When we
do emit subprogram DIEs, they do not contain any frame info or
parameter/variable children (which hugely reduces the number of types
in .debug_info).

This version has the compiler/assembler changes; linker changes needed
will come in a subsequent patch.

Change-Id: I8ed404f94ca9847e5b97593097d6a3de63d4da96