cmd/link: add support for "minimal" DWARF gen
DO NOT SUBMIT
DO NOT REVIEW

This adds support to the linker for "minimal" DWARF generation (e.g.
line table plus pruned-down subprogram/inline entries in debug_info).

This implementation of the linker support doesn't include an explicit
linker command line option; instead the linker just keys off what it
sees from the compiler. That is to say, if the compiler emits "full"
debug info for a compile unit the the linker will add just that
compile unit's DWARF to the appropriate sections.

There are a couple of areas where it is less clear:

  - generation of key internal types (e.g. "runtime._type")

  - generation of .debug_frame (which currently is monolithic
    and does not key off what's happening in individual compile units)

Includes an addition to one of the linker's DWARF tests to ensure that
things are indeed missing (as expected) from the DWARF when
-minimaldwarf is in effect.

Change-Id: I98eaf87fbc3883ae7e9ff67740297e44f8172d9a