cmd/compile: make ppc64 rules less picky about type of 64-bit stores
all not-ppc64 architectures rewrite word-sized stores with the cases
"float, not-a-float", where single-word interface data is "not a float".
Ppc64 instead attempted to enumerate all the not-a-float types, which
did not include single-word-interface types.

OpArg retains its declared type, which can be a single-word interface,
so this can cause a problem.

The repaired code may be technically incorrect (it is a dubious type
pun that nonetheless works on all other architectures, so far) but it
is now uniformly technically incorrect.

Fixes #43551.

Change-Id: Ie0d91fc5b151db4cbb14aaf6d3ac66490e0228b7