cmd/{compile,link}: relocate generation of DWARF for global vars
Move DWARF generation for global variables from the linker to the
compiler. This effectively parallelizes this part of DWARF generation,
speeds up the linker minutely, and gives us a slightly more rational
implementation (there was really no compelling reason to do DWARF gen
for globals in the linker).

Note: in this version of the patch the old linker-based implementation
is still present, and the new behavior is gated by compiler and linker
command line options. A subsequent patch will remove this.

Change-Id: I7476ebd9cfe02d7e089f56b20110b044ba5ce57f