cmd/internal/obj/arm64: fix VMOVQ instruction encoding error
The VMOVQ instruction moves a 128-bit constant into a V register, as 128-bit
constant can't be loaded into a register directly, we split it into two 64-bit
constants and load it from constant pool. So we need to call the 'addpool'
function twice, this function sets the 'Pool' field of a 'Prog', however if
we set this field twice for the same Prog, the PC-relative offset will be wrong.

Besides, the CL deletes the check of the operand class in addpool, it seems
unnecessary, without this check we don't need to list all compatible classes.

Change-Id: I616f043c99a9a18a663f8768842cc980de2e6f79