cmd/compile: allow embed into any string or byte slice type
The current implementation requires saying "string" or "[]byte"
and disallows aliases, defined types, and even "[]uint8".
This was not 100% intended and mostly just fell out of when
the checks were being done in the implementation (too early,
before typechecking).

After discussion on #43217, the consensus was to allow
all string and byte slice types, same as we do for string
conversions in the language itself. This CL does that.

It's more code than you'd expect because the decision has
to be delayed until after typechecking.

But it also more closely aligns with the version that's
already on dev.regabi.

For #43217.

DO NOT SUBMIT - #43217 has not been accepted yet.

Change-Id: Iba919cfadfbd5d7116f2bf47e2512fb1d5c36731