cmd/compile: remove embed import restriction
The import restriction was a bit surprising when you
are just trying to embed into a []byte or string.

The original motivation was performance, to avoid
scanning the file for //go:embed lines during go build,
but we already read the full text of every file later
during go build anyway (to hash them to check for staleness),
so there's not much harm in reading them earlier too.

After discussion on #43217, the consensus was to remove
the import restriction.

Fixes #43217.

DO NOT SUBMIT - #43217 has not been accepted yet.

Change-Id: Icb8f69a99e7883d09eaf608acb63f06616ed3c98