[dev.fuzz] internal/fuzz: fix context cancellation for coordinator
Previously, performing a SIGINT would cause the coordinator
to write a crash to testdata, and would continue to run
despite being interupted.

Also includes a few small cleanups.

Note that this still only works with -parallel=1, and does
not correctly cancel all processes when more than one worker
is running.

Change-Id: Ia3cf7cd231c30ac9ad2a61f4935aa543e241f60d