[dev.fuzz] internal/fuzz: add a lock for accessing shared memory
We are experiencing transient segfaults when calling
setValue in *workerClient.fuzz(). One possible reason
this could occur would be if the worker resources,
including the shared memory file, get cleaned and removed
while the fuzzing go routine is still running. To avoid
this, the shared memory region is now protected by a
mutex.

Change-Id: I906462982e8ecac51d1074f21f6357834be5be42