[dev.regabi] cmd/compile: remove FuncInfo.StackObjects
By simply calling objw.Global and x.Set right away, like we do in
WriteFuncMap, we can get rid of having to track the stkobj symbol
separately. It only needs to be referenced via the FUNCDATA.

Does not pass toolstash -cmp, because it causes symbols to emitted in
a different order (due to the moved objw.Global).

Change-Id: Id6e2d4db8e221bc55679ea7e150500d88565b4d9