[dev.typeparams] cmd/compile: fix MethodExpr handling with embedded fields
The recent refactoring of SelectorExpr code to helpers broke the
handling of MethodExprs when there is an embedded field involved (e.g.
test/method7.go, line 48). If there is an embedded field involved, the
node op seen in DotMethod() is an ODOT rather than an OTYPE. Also, the
receiver type of the result should be the original type, but the new
code was using the last type after following the embedding path.

We could add a helper DotMethodExpr(), but it would need to take a type
argument as well, i.e. n := DotMethodExpr(typ, pos, x, last, typ)

Change-Id: I13f7ea6448b03d3e8f974103ee3a027219ca8388