cmd/vendor: add windows/arm64 support
DO NOT REVIEW / DO NOT SUBMIT

This change needs to be added upstream to
golang.org/x/sys/windows, and then this CL
needs to be redone as bringing in the new copy.

But for now this is enough to make the rest of
std cmd happy.

This CL is part of a stack adding windows/arm64
support (#36439), intended to land in the Go 1.17 cycle.

Change-Id: Ifaca710093376c658ecf91239aa05cc33af98a31