runtime: fix windows/arm CONTEXT_CONTROL
The constant was wrong, and the “right” constant doesn't work either.
But with the actually-right constant, profiling now works.

DO NOT REVIEW / DO NOT SUBMIT

I have not been able to test this on a windows/arm device.
However, the assertion here that the “real” CONTEXT_CONTROL
is insufficient for getting at LR is true of windows/arm64,
and profiling is known to be broken on windows/arm, so this
seems a likely candidate. But obviously this needs to be tested,
and this CL should reenable the pprof test if it does in fact fix it
("broken = true" should be removed for windows/arm).

Change-Id: If8caff1da556826db40961fb9bcfe2b1f31ea9f9